fix: 修改自定义函数报错
