+++ frontend/src/business/components/settings/project/function/custom-function.js
@@ -1,0 +1,7 @@
+export const FUNC_TEMPLATE = {
+  beanshell: "public static void test() {\n\n\n}",
+  groovy: "public static void test() {\n\n\n}",
+  python: "def test():\n",
+  nashornScript: "function test() {\n\n\n}",
+  rhinoScript: "function test() {\n\n\n}"
+}
